requirements.txt
python-dotenv
langchain_core 
langchain 
langchain-google-genai
streamlit
